requirements.txt

numpy<2
torch>=2.1

lightning>=2.1

PyYAML>=6.0

torchmetrics>=1.2
pandas

matplotlib